Update to use JSON remote commands
