+++ Cartfile.resolved
@@ -1,3 +1,2 @@
-binary "https://raw.githubusercontent.com/Appboy/appboy-ios-sdk/master/appboy_ios_sdk_core.json" "3.21.3"
-github "tealium/plcrashreporter" "1.3.7"
-github "tealium/tealium-swift" "1.9.3"
+binary "https://raw.githubusercontent.com/Appboy/appboy-ios-sdk/master/appboy_ios_sdk_core.json" "3.28.0"
+git "file:///Users/christina/Projects/work-repos/github/tealium-swift-builder" "5779423f239c6e21cf091571aa82a4f0b27e5177"
