+++ TealiumBraze.podspec
@@ -3,7 +3,7 @@ Pod::Spec.new do |s|
     # ―――  Spec Metadata  ―――――――――――――――――――――――――――――――――――――――――――――――――――――――――― #
     s.name         = "TealiumBraze"
     s.module_name  = "TealiumBraze"
-    s.version      = "1.0.0"
+    s.version      = "2.0.0"
     s.summary      = "Tealium Swift and Braze integration"
     s.description  = <<-DESC
     Tealium's integration with Braze for iOS.
@@ -20,7 +20,7 @@ Pod::Spec.new do |s|
     s.social_media_url   = "https://twitter.com/tealium"
 
     # ――― Platform Specifics ――――――――――――――――――――――――――――――――――――――――――――――――――――――― #
-    s.swift_version = "4.2"
+    s.swift_version = "5.0"
     s.platform     = :ios, "10.0"
     s.ios.deployment_target = "10.0"    
 
@@ -31,10 +31,9 @@ Pod::Spec.new do |s|
     s.ios.source_files      = "Sources/*.{swift}"
 
     # ――― Dependencies ―――――――――――――――――――――――――――――――――――――――――――――――――――――――――――――― #
-    s.ios.dependency 'tealium-swift/Core', '~> 1.9'
-    s.ios.dependency 'tealium-swift/TealiumRemoteCommands', '~> 1.9'
-    s.ios.dependency 'tealium-swift/TealiumDelegate', '~> 1.9'
-    s.ios.dependency 'tealium-swift/TealiumTagManagement', '~> 1.9'
-    s.ios.dependency 'Appboy-iOS-SDK', '~> 3.21'
+    s.ios.dependency 'tealium-swift/Core', '~> 2.0'
+    s.ios.dependency 'tealium-swift/RemoteCommands', '~> 2.0'
+    s.ios.dependency 'tealium-swift/TagManagement', '~> 2.0'
+    s.ios.dependency 'Appboy-iOS-SDK', '~> 3.2'
 
 end
