+++ TealiumBrazeExample/TealiumBrazeExample/EngagementViewController.swift
@@ -1,9 +1,9 @@
 //
 //  ViewController.swift
-//  BrazeRemoteCommand
+//  TealiumBrazeExample
 //
 //  Created by Jonathan Wong on 5/29/19.
-//  Copyright © 2019 Jonathan Wong. All rights reserved.
+//  Copyright © 2019 Tealium. All rights reserved.
 //
 
 import UIKit
@@ -11,6 +11,7 @@ import UIKit
 class EngagementViewController: UIViewController {
 
     @IBOutlet weak var twitterIdTextField: UITextField!
+    @IBOutlet weak var twitterNameTextField: UITextField!
     @IBOutlet weak var facebookUserTextField: UITextField!
     @IBOutlet weak var facebookFriendsTextField: UITextField!
     @IBOutlet weak var emailSegmentedControl: UISegmentedControl!
@@ -33,15 +34,33 @@ class EngagementViewController: UIViewController {
         }
         data["facebook_user"] = facebookInfo
         
-        data["twitter_id"] = "fattywaffles"
+        data["username"] = facebookUserTextField.text ?? ""
+        data["twitter_id"] = twitterIdTextField.text ?? ""
         data["twitter_user_description"] = "awesome engineer"
-        
+        data["twitter_name"] = twitterNameTextField.text ?? ""
+        data["list_of_likes"] = [
+            [
+              "name": "Bill the Cat",
+              "id": "155111347875779",
+              "created_time": "2017-06-18T18:21:04+0000"
+            ],
+            [
+              "name": "Calvin and Hobbes",
+              "id": "257573197608192",
+              "created_time": "2017-06-18T18:21:02+0000"
+            ],
+            [
+              "name": "Berkeley Breathed's Bloom County",
+              "id": "108793262484769",
+              "created_time": "2017-06-18T18:20:58+0000"
+            ]
+          ]
         data["email_subscription"] = emailSubscription()
         data["push_subscription"] = pushSubscription()
         
-        TealiumHelper.track(title: "setengagement", data: data)
-        TealiumHelper.track(title: "facebook", data: data)
-        TealiumHelper.track(title: "twitter", data: data)
+        TealiumHelper.trackEvent(title: "setengagement", data: data)
+        TealiumHelper.trackEvent(title: "facebook", data: data)
+        TealiumHelper.trackEvent(title: "twitter", data: data)
     }
     
     private func emailSubscription() -> String {
