+++ TealiumBrazeExample/TealiumBrazeExample/TealiumHelper.swift
@@ -1,49 +1,65 @@
 //
 //  TealiumHelper.swift
-//  BrazeRemoteCommand
+//  TealiumBrazeExample
 //
 //  Created by Jonathan Wong on 5/29/19.
-//  Copyright © 2019 Jonathan Wong. All rights reserved.
+//  Copyright © 2019 Tealium. All rights reserved.
 //
 
 import Foundation
 import TealiumSwift
 import TealiumBraze
 
+enum TealiumConfiguration {
+    static let account = "tealiummobile"
+    static let profile = "firebase-tag"
+    static let environment = "dev"
+}
+
 class TealiumHelper {
-    
+
     static let shared = TealiumHelper()
-    let config = TealiumConfig(account: "tealiummobile",
-                               profile: "braze-tag",
-                               environment: "qa",
-                               datasource: nil)
-    
+
+    let config = TealiumConfig(account: TealiumConfiguration.account,
+        profile: TealiumConfiguration.profile,
+        environment: TealiumConfiguration.environment)
+
     var tealium: Tealium?
+    
     static var universalData = [String: Any]()
     
+    // JSON Remote Command
+    let brazeRemoteCommand = BrazeRemoteCommand(type: .remote(url: "https://tags.tiqcdn.com/dle/tealiummobile/demo/braze.json"))
+
     private init() {
-        config.logLevel = .verbose
         config.shouldUseRemotePublishSettings = false
+        config.batchingEnabled = false
+        config.remoteAPIEnabled = true
+        config.logLevel = .info
+        config.collectors = [Collectors.Lifecycle]
+        config.dispatchers = [Dispatchers.TagManagement, Dispatchers.RemoteCommands]
         
-        tealium = Tealium(config: config) { responses in
-            guard let remoteCommands = self.tealium?.remoteCommands() else {
-                return
-            }
-            let brazeTracker = BrazeTracker()
-            let brazeCommand = BrazeCommand(brazeTracker: brazeTracker)
-            let brazeRemoteCommand = brazeCommand.remoteCommand()
-            remoteCommands.add(brazeRemoteCommand)
-        }
+        config.addRemoteCommand(brazeRemoteCommand)
+        
+        tealium = Tealium(config: config)
     }
-    
-    class func start() {
+
+
+    public func start() {
         _ = TealiumHelper.shared
     }
-    
-    class func track(title: String, data: [String: Any]?) {
+
+    class func trackView(title: String, data: [String: Any]?) {
+        let tealiumView = TealiumView(title, dataLayer: data)
+        TealiumHelper.shared.tealium?.track(tealiumView)
+    }
+
+    class func trackEvent(title: String, data: [String: Any]?) {
         if let data = data {
             universalData = universalData.merging(data) { _, new in new }
         }
-        TealiumHelper.shared.tealium?.track(title: title, data: universalData, completion: nil)
+        let tealiumEvent = TealiumEvent(title, dataLayer: universalData)
+        TealiumHelper.shared.tealium?.track(tealiumEvent)
     }
+
 }
